- refactoring - fix created_at field
