+++ schema/date.js
@@ -1,0 +1,25 @@
+const graphql = require('graphql');
+const {
+    GraphQLScalarType
+} = graphql;
+const { Kind } = require('graphql/language');
+
+
+const Date = new GraphQLScalarType({
+    name: 'Date',
+    description: 'Date custom scalar type',
+    parseValue(value) {
+        return new Date(value); // value from the client
+    },
+    serialize(value) {
+        return value.getTime(); // value sent to the client
+    },
+    parseLiteral(ast) {
+        if (ast.kind === Kind.INT) {
+            return new Date(ast.value) // ast value is always in string format
+        }
+        return null;
+    },
+})
+
+module.exports = Date;
